Added a double verion of cellType since GPUGridVariables only support doubles at this time. This hack creates polution but should not influence the results.
The CPU & GPU code is heavily instrumented for direct comparisons.

+++ CCA/Components/Models/Radiation/RMCRT/RayGPU.cuh
@@ -143,7 +143,7 @@ __device__ void updateSumIDevice ( double3& ray_direction,
                                    const double3& Dx,
                                    const GPUGridVariable<double>&  sigmaT4OverPi,
                                    const GPUGridVariable<double>& abskg,
-                                   const GPUGridVariable<int>& celltype,
+                                   const GPUGridVariable<double>& celltype,
                                    double& sumI,
                                    curandState* randNumStates,
                                    RMCRT_flags RT_flags);
